auc-28 refactor : infra 패키지명 오타 수정
